+++ backend/borg/delete.go
