+++ backend/borg/info.go
